requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==4.2.22
django-debug-toolbar==5.2.0
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
django-extensions
djangorestframework
requests
drf-yasg
django-cors-headers
